Changed loginUser funtion to file login-controller
